move all API calls from utiltities to API call file
