+++ src/App/App.js
@@ -4,7 +4,7 @@ import Name from '../Name/Name'
 import Question from '../Question/Question'
 import Result from '../Result/Result'
 import Error from '../Error/Error'
-import { getHogwartsHouses } from '../utilities'
+import { getHogwartsHouses } from '../apiCalls'
 import React, { Component } from 'react'
 import { Route, Switch } from 'react-router-dom'
